Moving unit converters to tools directory
