+++ parcels/tools/__init__.py
@@ -1,1 +1,2 @@
-from . import timer  # noqa
+from .timer import *  # noqa
+from .converters import *  # noqa
